requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.2
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
idna==3.9
jose==1.0.0
magic-filter==1.0.12
multidict==6.1.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.11.1
